Refactor KSailUpCommandTests and add E2E tests for various configurations
